Split synchronous tests of test_worker into a dedicated file
